refactor: todoListApp 함수 생성 및 index.js에서 분리
